+++ src/main/java/insane96mcp/survivalreimagined/mixin/AbstractArrowMixin.java
@@ -47,6 +47,7 @@ public abstract class AbstractArrowMixin extends Projectile {
         return this.isCritArrow();
     }
 
+    //Disable mobs arrow random bonus damage
     @Inject(method = "setEnchantmentEffectsFromEntity", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/projectile/AbstractArrow;setBaseDamage(D)V", ordinal = 0, shift = At.Shift.AFTER))
     private void onSetMobArrowDamage(LivingEntity pShooter, float pVelocity, CallbackInfo ci) {
         this.setBaseDamage(pVelocity * 2.0F);