+++ changelog.md
@@ -1,12 +1,12 @@
 # Changelog
 
-* Increase bell range when finding riders
 * Reduce arrows base damage?
 
 ## Upcoming
 * Food can now take 1.2 seconds to eat if low saturation and hunger enough (was 1.6) (fast food like kelp still doesn't go below 0.8 secs)
 * Ensorceller now gets power from nearby sunflowers, making lower level rolls available earlier and later in the day
 * Doubled the bell range to glow raiders
+* Reduced arrows from bows base damage
 * Fixed MC-152728
 * Fixed ensorceller